rename $package_name to $distribution_name
It never really specified the name of a *package*, but instead
it specifies the name of the distribution archive and the name
of the folder in which the distribution will live.

This also makes it possible to implement *actual* package
installations in the future on platforms where SonarQube packages
are available, i.e. FreeBSD.